--- CHANGELOG.md
@@ -1,84 +1,0 @@
-# Changelog
-
-## 1.11.4 - 2024-04-18
-
-- Replaced Makefile with GoReleaser
-- Added Dependabot
-- Added GitHub Actions workflows
-
-## 1.11.3 - 2024-03-30
-
-- Added macOS support
-
-## 1.11.2 - 2023-08-01
-
-- Updated version text
-
-## 1.11.1 - 2023-07-09
-
-- Dockerbuild to 1.20
-
-## 1.11.0 - 2023-07-09
-
-- Upgrade to Go 1.20
-- Windows ARM64 binary
-- Devcontainer improvements
-- Added Dependabot config
-
-## 1.10.0 - 2022-11-02
-
-- Upgrade to Go 1.17
-- Default DEB compression
-
-## 1.9.1 - 2021-07-04
-
-- Debian ARM64 package
-- Go version numbering
-
-## 1.9 - 2021-07-03
-
-- Updated dependencies
-- FreeBSD and Linux arm64 support
-
-## 1.8 - 2021-04-16
-
-Mostly technical changes:
-- OpenBSD support
-- GitPod
-- man page
-
-## 1.7 - 2019-10-01
-
-### Added
-
-### Changed
-
-- `xclip` is a dependency of the Debian package
-- Switched to Go modules
-- Fixed editor which requires tty
-- Added FreeBSD build
-
-### Removed
-
-## 1.6 - 2019-03-01
-
-### Added
-
-- Debian packaging
-
-### Changed
-
-- Fix for showing wrong section
-
-### Removed
-
-## 1.5 - 2019-02-06
-
-### Added
-
-- `-q` option for no output
-- Makefile for building
-
-### Changed
-
-### Removed
