+++ internal/csheet/reader.go
@@ -1,8 +1,7 @@
-package main
+package csheet
 
 import (
 	"bufio"
-	"flag"
 	"fmt"
 	"io"
 	"os"
@@ -11,39 +10,6 @@ import (
 
 	"github.com/atotto/clipboard"
 )
-
-var csheetFile string
-var version = "DEV-BUILD"
-
-func main() {
-	var clipboardArg = flag.Bool("c", false, "Copy result to clipboard")
-	var fileArg = flag.String("f", "", "Cheat sheet Mardown file")
-	var listArg = flag.Bool("l", false, "Show all possible entries")
-	var quietArg = flag.Bool("q", false, "No output")
-	var versionArg = flag.Bool("v", false, "Display version")
-
-	flag.Parse()
-
-	args := flag.Args()
-	validateArgs(fileArg, listArg, versionArg, args)
-
-	if *fileArg == "" {
-		csheetFile = getCSheetDir() + string(os.PathSeparator) + "csheet.md"
-	} else {
-		csheetFile = *fileArg
-	}
-
-	if *versionArg {
-		printVersion()
-	} else if *listArg {
-		printEntries(*clipboardArg, *quietArg)
-	} else {
-		subject := args[0]
-		section := args[1]
-
-		printEntry(subject, section, *clipboardArg, *quietArg)
-	}
-}
 
 func findEntry(fp *os.File, subject string, section string) []string {
 	r := bufio.NewReaderSize(fp, 4*1024)
@@ -94,7 +60,7 @@ func findHeader(r *bufio.Reader, header string, demarcation *string) bool {
 	return false
 }
 
-func getCSheetDir() string {
+func GetCSheetDir() string {
 	usr, err := user.Current()
 	if err != nil {
 		panic(err)
@@ -103,7 +69,7 @@ func getCSheetDir() string {
 	return usr.HomeDir
 }
 
-func openFile() (fp *os.File) {
+func openFile(csheetFile string) (fp *os.File) {
 	_, err := os.Stat(csheetFile)
 	if err == nil {
 		fp, err := os.Open(csheetFile)
@@ -124,8 +90,8 @@ func openFile() (fp *os.File) {
 	}
 }
 
-func printEntry(subject string, section string, copyToClipboard bool, quiet bool) {
-	fp := openFile()
+func PrintEntry(csheetFile string, subject string, section string, copyToClipboard bool, quiet bool) {
+	fp := openFile(csheetFile)
 	defer fp.Close()
 
 	code := findEntry(fp, subject, section)
@@ -142,8 +108,8 @@ func printEntry(subject string, section string, copyToClipboard bool, quiet bool
 	}
 }
 
-func printEntries(copyToClipboard bool, quiet bool) {
-	fp := openFile()
+func PrintEntries(csheetFile string, copyToClipboard bool, quiet bool) {
+	fp := openFile(csheetFile)
 	defer fp.Close()
 
 	entries := findEntries(fp)
@@ -158,29 +124,6 @@ func printEntries(copyToClipboard bool, quiet bool) {
 		clipboardEntries := strings.Join(entries, "\n")
 		clipboard.WriteAll(clipboardEntries)
 	}
-}
-
-func printUsage() {
-	fmt.Println("Usage: csheet { OPTIONS } [SUBJECT] [SECTION]")
-	fmt.Println("Options:")
-	fmt.Println("-c        : Copy result to clipboard")
-	fmt.Println("-f [FILE] : Specifies the Markdown file to read")
-	fmt.Println("-h        : Print help")
-	fmt.Println("-l        : Show all possible entries")
-	fmt.Println("-q        : No output, useful with -c")
-	fmt.Println("-v        : Shows the versions")
-}
-
-func printVersion() {
-	versionInfo := strings.Split(version, "-")
-	csheetVersion := strings.Join(versionInfo[:len(versionInfo)-1], "-")
-	csheetRevision := versionInfo[len(versionInfo)-1]
-
-	fmt.Printf("csheet version v%s, revision %s", csheetVersion, csheetRevision)
-	fmt.Println("")
-	fmt.Println("See: https://github.com/ninckblokje/csheet")
-	fmt.Println("")
-	fmt.Println("For my kids, L&M")
 }
 
 func readCode(r *bufio.Reader) []string {
@@ -221,22 +164,6 @@ func readLine(r *bufio.Reader) *string {
 	}
 
 	return nil
-}
-
-func validateArgs(fileArg *string, listArg *bool, versionArg *bool, args []string) {
-	if *versionArg {
-		// ok
-		return
-	} else if *listArg {
-		// ok
-		return
-	} else if len(args) == 2 {
-		// ok
-		return
-	} else {
-		printUsage()
-		os.Exit(1)
-	}
 }
 
 func writeHeader(fp *os.File, header string) {
